handle Lmax as double instead of long

+++ api/src/main/java/au/org/aodn/nrmn/restapi/model/db/UiSpeciesAttributes.java
@@ -15,6 +15,6 @@ public interface UiSpeciesAttributes {
 
      Long getMaxAbundance();
 
-     Long getLmax();
+     Double getLmax();
 
 }
